rebuild.sh: Save dependencies in build directory.
This patch ensures that if the --build-dir=<path> option is being
used, the compiled dependency libraries (zlib, glib, etc...) are
kept and reused on each invokation of the rebuild.sh script.

This saves a considerable amount of time when performing incremental
compilation, i.e. when hacking on the qemu-android sources.